Let users to skip project and bucket creation
Users can now supply a project ID and a bucket name to store the
Terraform state without provisioning them using the Terraform
descriptors in the terraform-init directory. This enables users to avoid
creating a project and a storage bucket if the have other means and
processes to do so, instead of relying on the Terraform descriptors we
provide to do so.

This behavior is disabled by default, and users can enable it by passing
the -n (--no-provision-google-cloud-project) option to the
provision-cloud-infrastructure.sh script.

Change-Id: Ifb778f5fcafbeb2fa55620f1bd225df3a33e631b